Remove zPIV code from main.cpp
Github-Pull: #609
Rebased-From: 873ef19fc0dd754bc127ff6f94a2ddec650f06d7
