+++ src/blocksignature.cpp
@@ -4,6 +4,7 @@
 
 #include "blocksignature.h"
 #include "main.h"
+#include "zpivchain.h"
 
 bool SignBlockWithKey(CBlock& block, const CKey& key)
 {
