+++ src/qt/transactionrecord.cpp
@@ -11,6 +11,7 @@
 #include "swifttx.h"
 #include "timedata.h"
 #include "wallet.h"
+#include "zpivchain.h"
 
 #include <stdint.h>
