+++ src/rpcwallet.cpp
@@ -17,6 +17,7 @@
 #include "utilmoneystr.h"
 #include "wallet.h"
 #include "walletdb.h"
+#include "zpivchain.h"
 
 #include <stdint.h>
