+++ src/wallet.cpp
@@ -24,6 +24,7 @@
 #include "txdb.h"
 #include "util.h"
 #include "utilmoneystr.h"
+#include "zpivchain.h"
 
 #include "denomination_functions.h"
 #include "libzerocoin/Denominations.h"
